server: fix registration edge case check when the address is already registered
this is an edge case - if the wallet is already registered and say the node
is moved to some other machine or if the repo is deleted and started all over again
then there wouldn't be any file at ../node-lite/register-tx-hash.txt because contracts/lite-register.js
would return from `if (await checkIfRegistered(signerAddress))` condition and would not write anything
to txHashPath in `fs.writeFileSync(txHashPath, tx.hash);`. So now the server code checks if the
signer address is already registered and thus prevents the error caused in `node-lite/register.js`'s
`Step:5 console.log(orange.bold('\n🔶 Step 5: Server Registration'))`.

Signed-off-by: 0xnightwind <nightwind1899@gmail.com>